+++ pycontrails/models/cocipgrid/cocip_grid.py
@@ -1102,21 +1102,38 @@ def _evolve_vector(
         dt_head = dt - half_head_tail_dt  # type: ignore[operator]
         dt_tail = dt + half_head_tail_dt  # type: ignore[operator]
 
-    # After advection, out has time t
-    out = advect(vector, dt, dt_head, dt_tail)  # type: ignore[arg-type]
+    vector_1 = vector
+    vector_12 = vector
+    vector_2 = advect(vector_1, vector_12, dt, dt_head, dt_tail)  # type: ignore[arg-type]
 
-    out = run_interpolators(
-        out,
+    vector_2 = run_interpolators(
+        vector_2,
         met,
         rad,
         dz_m=params["dz_m"],
         humidity_scaling=params["humidity_scaling"],
         **params["_interp_kwargs"],
     )
-    out = calc_evolve_one_step(vector, out, params)
-    ef_summary = out.select(("index", "age", "ef"), copy=False)
 
-    return out, ef_summary
+    vector_2, persistent = calc_evolve_one_step(vector, vector_12, vector_2, params)
+
+    if params["second_order_runge"]:
+        vector_12 = vector_2
+        vector_2 = advect(vector_1, vector_12, dt, dt_head, dt_tail)  # type: ignore[arg-type]
+
+        vector_2 = run_interpolators(
+            vector_2,
+            met,
+            rad,
+            dz_m=params["dz_m"],
+            humidity_scaling=params["humidity_scaling"],
+            **params["_interp_kwargs"],
+        )
+        vector_2, persistent = calc_evolve_one_step(vector, vector_12, vector_2, params)
+
+    vector_2 = vector_2.filter(persistent)
+    ef_summary = vector_2.select(("index", "age", "ef"), copy=False)
+    return vector_2, ef_summary
 
 
 def _run_downwash(
@@ -1585,144 +1602,152 @@ def find_initial_persistent_contrails(
 
 
 def calc_evolve_one_step(
-    curr_contrail: GeoVectorDataset,
-    next_contrail: GeoVectorDataset,
+    contrail_1: GeoVectorDataset,
+    contrail_12: GeoVectorDataset,
+    contrail_2: GeoVectorDataset,
     params: dict[str, Any],
-) -> GeoVectorDataset:
-    """Calculate contrail properties of ``next_contrail``.
+) -> tuple[GeoVectorDataset, npt.NDArray[np.bool_]]:
+    """Calculate contrail properties of ``contrail_2``.
 
-    This function attaches additional variables to ``next_contrail``, then
-    filters by :func:`contrail_properties.contrail_persistent`.
+    This function attaches additional variables to ``contrail_2`` (it is
+    freely modified in place). The function also returns a boolean array of
+    filters produced by :func:`contrail_properties.contrail_persistent`.
+
+    This implementation parallels
+    :func:pycontrails.models.cocip.cocip.time_integration_runge_kutta`. In particular,
+    the same conventions and variable names are used when possible.
 
     Parameters
     ----------
-    curr_contrail : GeoVectorDataset
+    contrail_1 : GeoVectorDataset
         Existing contrail
-    next_contrail : GeoVectorDataset
+    contrail_12 : GeoVectorDataset
+        Intermediate contrail used for the Runge-Kutta method
+    contrail_2 : GeoVectorDataset
         Result of advecting existing contrail already interpolated against CoCiP met data
     params : dict[str, Any]
         CoCiP model parameters. See :class:`CocipGrid`.
 
     Returns
     -------
-    GeoVectorDataset
-        Parameter ``next_contrail`` filtered by persistence.
+    tuple[GeoVectorDataset, npt.NDArray[np.bool_]]
+        A tuple with ``contrail_2`` and a boolean array indicating which
+        points persist after the evolution step.
     """
     calc_wind_shear(
-        next_contrail,
+        contrail_2,
         is_downwash=False,
         dz_m=params["dz_m"],
         dsn_dz_factor=params["dsn_dz_factor"],
     )
-    calc_thermal_properties(next_contrail)
+    calc_thermal_properties(contrail_2)
 
-    iwc_t1 = curr_contrail["iwc"]
-    specific_humidity_t1 = curr_contrail["specific_humidity"]
-    specific_humidity_t2 = next_contrail["specific_humidity"]
-    q_sat_t1 = curr_contrail["q_sat"]
-    q_sat_t2 = next_contrail["q_sat"]
-    plume_mass_per_m_t1 = curr_contrail["plume_mass_per_m"]
-    width_t1 = curr_contrail["width"]
-    depth_t1 = curr_contrail["depth"]
-    sigma_yz_t1 = curr_contrail["sigma_yz"]
-    dsn_dz_t1 = curr_contrail["dsn_dz"]
-    diffuse_h_t1 = curr_contrail["diffuse_h"]
-    diffuse_v_t1 = curr_contrail["diffuse_v"]
+    iwc_1 = contrail_1["iwc"]
+    specific_humidity_1 = contrail_1["specific_humidity"]
+    specific_humidity_2 = contrail_2["specific_humidity"]
+    q_sat_1 = contrail_1["q_sat"]
+    q_sat_2 = contrail_2["q_sat"]
+    plume_mass_per_m_1 = contrail_1["plume_mass_per_m"]
+    width_1 = contrail_1["width"]
+    depth_1 = contrail_1["depth"]
+    sigma_yz_1 = contrail_1["sigma_yz"]
+
+    dsn_dz_12 = contrail_12["dsn_dz"]
+    diffuse_h_12 = contrail_12["diffuse_h"]
+    diffuse_v_12 = contrail_12["diffuse_v"]
 
     # Segment-free mode logic
-    segment_length_t2: np.ndarray | float
-    seg_ratio_t12: np.ndarray | float
-    if _is_segment_free_mode(curr_contrail):
-        segment_length_t2 = 1.0
-        seg_ratio_t12 = 1.0
+    segment_length_2: np.ndarray | float
+    seg_ratio_12: np.ndarray | float
+    if _is_segment_free_mode(contrail_1):
+        segment_length_2 = 1.0
+        seg_ratio_12 = 1.0
     else:
-        segment_length_t1 = curr_contrail["segment_length"]
-        segment_length_t2 = next_contrail["segment_length"]
-        seg_ratio_t12 = contrail_properties.segment_length_ratio(
-            segment_length_t1, segment_length_t2
-        )
+        segment_length_1 = contrail_1["segment_length"]
+        segment_length_2 = contrail_2["segment_length"]
+        seg_ratio_12 = contrail_properties.segment_length_ratio(segment_length_1, segment_length_2)
 
-    dt = next_contrail["time"] - curr_contrail["time"]
+    dt = contrail_2["time"] - contrail_1["time"]
 
-    sigma_yy_t2, sigma_zz_t2, sigma_yz_t2 = contrail_properties.plume_temporal_evolution(
-        width_t1=width_t1,
-        depth_t1=depth_t1,
-        sigma_yz_t1=sigma_yz_t1,
-        dsn_dz_t1=dsn_dz_t1,
-        diffuse_h_t1=diffuse_h_t1,
-        diffuse_v_t1=diffuse_v_t1,
-        seg_ratio=seg_ratio_t12,
+    sigma_yy_2, sigma_zz_2, sigma_yz_2 = contrail_properties.plume_temporal_evolution(
+        width_t1=width_1,
+        depth_t1=depth_1,
+        sigma_yz_t1=sigma_yz_1,
+        dsn_dz_t1=dsn_dz_12,
+        diffuse_h_t1=diffuse_h_12,
+        diffuse_v_t1=diffuse_v_12,
+        seg_ratio=seg_ratio_12,
         dt=dt,
         max_depth=params["max_depth"],
     )
 
-    width_t2, depth_t2 = contrail_properties.new_contrail_dimensions(sigma_yy_t2, sigma_zz_t2)
-    next_contrail["sigma_yz"] = sigma_yz_t2
-    next_contrail["width"] = width_t2
-    next_contrail["depth"] = depth_t2
+    width_2, depth_2 = contrail_properties.new_contrail_dimensions(sigma_yy_2, sigma_zz_2)
+    contrail_2["sigma_yz"] = sigma_yz_2
+    contrail_2["width"] = width_2
+    contrail_2["depth"] = depth_2
 
-    area_eff_t2 = contrail_properties.new_effective_area_from_sigma(
-        sigma_yy=sigma_yy_t2,
-        sigma_zz=sigma_zz_t2,
-        sigma_yz=sigma_yz_t2,
+    area_eff_2 = contrail_properties.new_effective_area_from_sigma(
+        sigma_yy=sigma_yy_2,
+        sigma_zz=sigma_zz_2,
+        sigma_yz=sigma_yz_2,
     )
 
-    rho_air_t2 = next_contrail["rho_air"]
-    plume_mass_per_m_t2 = contrail_properties.plume_mass_per_distance(area_eff_t2, rho_air_t2)
-    iwc_t2 = contrail_properties.new_ice_water_content(
-        iwc_t1=iwc_t1,
-        q_t1=specific_humidity_t1,
-        q_t2=specific_humidity_t2,
-        q_sat_t1=q_sat_t1,
-        q_sat_t2=q_sat_t2,
-        mass_plume_t1=plume_mass_per_m_t1,
-        mass_plume_t2=plume_mass_per_m_t2,
+    rho_air_2 = contrail_2["rho_air"]
+    plume_mass_per_m_2 = contrail_properties.plume_mass_per_distance(area_eff_2, rho_air_2)
+    iwc_2 = contrail_properties.new_ice_water_content(
+        iwc_t1=iwc_1,
+        q_t1=specific_humidity_1,
+        q_t2=specific_humidity_2,
+        q_sat_t1=q_sat_1,
+        q_sat_t2=q_sat_2,
+        mass_plume_t1=plume_mass_per_m_1,
+        mass_plume_t2=plume_mass_per_m_2,
     )
-    next_contrail["iwc"] = iwc_t2
+    contrail_2["iwc"] = iwc_2
 
-    n_ice_per_m_t1 = curr_contrail["n_ice_per_m"]
-    dn_dt_agg = curr_contrail["dn_dt_agg"]
-    dn_dt_turb = curr_contrail["dn_dt_turb"]
+    n_ice_per_m_t1 = contrail_1["n_ice_per_m"]
+    dn_dt_agg_12 = contrail_12["dn_dt_agg"]
+    dn_dt_turb_12 = contrail_12["dn_dt_turb"]
 
     n_ice_per_m_t2 = contrail_properties.new_ice_particle_number(
         n_ice_per_m_t1=n_ice_per_m_t1,
-        dn_dt_agg=dn_dt_agg,
-        dn_dt_turb=dn_dt_turb,
-        seg_ratio=seg_ratio_t12,
+        dn_dt_agg=dn_dt_agg_12,
+        dn_dt_turb=dn_dt_turb_12,
+        seg_ratio=seg_ratio_12,
         dt=dt,
     )
-    next_contrail["n_ice_per_m"] = n_ice_per_m_t2
+    contrail_2["n_ice_per_m"] = n_ice_per_m_t2
 
     cocip.calc_contrail_properties(
-        next_contrail,
+        contrail_2,
         params["effective_vertical_resolution"],
         params["wind_shear_enhancement_exponent"],
         params["sedimentation_impact_factor"],
         radiative_heating_effects=False,  # Not yet supported in CocipGrid
     )
-    cocip.calc_radiative_properties(next_contrail, params)
+    cocip.calc_radiative_properties(contrail_2, params)
 
-    rf_net_t1 = curr_contrail["rf_net"]
-    rf_net_t2 = next_contrail["rf_net"]
+    rf_net_t1 = contrail_1["rf_net"]
+    rf_net_t2 = contrail_2["rf_net"]
     ef = contrail_properties.energy_forcing(
         rf_net_t1=rf_net_t1,
         rf_net_t2=rf_net_t2,
-        width_t1=width_t1,
-        width_t2=width_t2,
-        seg_length_t2=segment_length_t2,
+        width_t1=width_1,
+        width_t2=width_2,
+        seg_length_t2=segment_length_2,
         dt=dt,
     )
     # NOTE: This will get masked below if `persistent` is False
     # That is, we are taking a right Riemann sum of a decreasing function, so we are
     # underestimating the truth. With dt small enough, this is fine.
-    next_contrail["ef"] = ef
+    contrail_2["ef"] = ef
 
-    # NOTE: Only dealing with `next_contrail` here
-    latitude = next_contrail["latitude"]
-    altitude = next_contrail["altitude"]
-    tau_contrail = next_contrail["tau_contrail"]
-    n_ice_per_vol = next_contrail["n_ice_per_vol"]
-    age = next_contrail["age"]
+    # NOTE: Only dealing with `contrail_2` here
+    latitude = contrail_2["latitude"]
+    altitude = contrail_2["altitude"]
+    tau_contrail = contrail_2["tau_contrail"]
+    n_ice_per_vol = contrail_2["n_ice_per_vol"]
+    age = contrail_2["age"]
 
     # Both tau_contrail and n_ice_per_vol could have nan values
     # These are mostly due to out of bounds interpolation
@@ -1734,23 +1759,14 @@ def calc_evolve_one_step(
     persistent = contrail_properties.contrail_persistent(
         latitude=latitude,
         altitude=altitude,
-        segment_length=segment_length_t2,  # type: ignore[arg-type]
+        segment_length=segment_length_2,  # type: ignore[arg-type]
         age=age,
         tau_contrail=tau_contrail,
         n_ice_per_m3=n_ice_per_vol,
         params=params,
     )
 
-    # Filter by persistent
-    logger.debug(
-        "Fraction of grid points surviving: %s / %s",
-        np.sum(persistent),
-        next_contrail.size,
-    )
-    if params["persistent_buffer"] is not None:
-        # See Cocip implementation if we want to support this
-        raise NotImplementedError
-    return next_contrail.filter(persistent)
+    return contrail_2, persistent
 
 
 def calc_emissions(vector: GeoVectorDataset, params: dict[str, Any]) -> None:
@@ -1962,25 +1978,32 @@ def calc_thermal_properties(contrail: GeoVectorDataset) -> None:
 
 
 def advect(
-    contrail: GeoVectorDataset,
+    contrail_1: GeoVectorDataset,
+    contrail_12: GeoVectorDataset,
     dt: np.timedelta64 | npt.NDArray[np.timedelta64],
     dt_head: np.timedelta64 | None,
     dt_tail: np.timedelta64 | None,
 ) -> GeoVectorDataset:
     """Form new contrail by advecting existing contrail.
 
-    Parameter ``contrail`` is not modified.
-
     .. versionchanged:: 0.25.0
 
-        The ``dt_head`` and ``dt_tail`` parameters are no longer optional.
+        The ``dt_head`` and ``dt_tail`` parameters are required.
         Set these to ``dt`` to evolve the contrail uniformly over a constant time.
         Set to None for segment-free mode.
 
+    .. versionchanged:: 0.54.4
+
+        Rename ``contrail`` parameter to ``contrail_1``.
+        Add ``contrail_12`` parameter.
+
     Parameters
     ----------
-    contrail : GeoVectorDataset
+    contrail_1 : GeoVectorDataset
         Grid points already interpolated against wind data
+    contrail_12 : GeoVectorDataset
+        Grid points with intermediate wind data.
+        See :func:`pycontrails.models.cocip.cocip.time_integration_runge_kutta`.
     dt : np.timedelta64 | npt.NDArray[np.timedelta64]
         Time step for advection
     dt_head : np.timedelta64 | None
@@ -2008,60 +2031,82 @@ def advect(
             - "segment_length"  (only if `is_segment_free=False`)
             - "head_tail_dt"  (only if `is_segment_free=False`)
     """
-    longitude = contrail["longitude"]
-    latitude = contrail["latitude"]
-    level = contrail["level"]
-    time = contrail["time"]
-    formation_time = contrail["formation_time"]
-    age = contrail["age"]
-    u_wind = contrail["eastward_wind"]
-    v_wind = contrail["northward_wind"]
-    vertical_velocity = contrail["lagrangian_tendency_of_air_pressure"]
-    rho_air = contrail["rho_air"]
-    terminal_fall_speed = contrail["terminal_fall_speed"]
+    longitude_1 = contrail_1["longitude"]
+    latitude_1 = contrail_1["latitude"]
+    level_1 = contrail_1["level"]
+    time_1 = contrail_1["time"]
+    age_1 = contrail_1["age"]
+    u_wind_1 = contrail_1["eastward_wind"]
+    v_wind_1 = contrail_1["northward_wind"]
+    vertical_velocity_1 = contrail_1["lagrangian_tendency_of_air_pressure"]
+    rho_air_1 = contrail_1["rho_air"]
 
-    # Using the _t2 convention for post-advection data
-    index_t2 = contrail["index"]
-    time_t2 = time + dt
-    age_t2 = age + dt
+    u_wind_12 = contrail_12["eastward_wind"]
+    v_wind_12 = contrail_12["northward_wind"]
+    vertical_velocity_12 = contrail_12["lagrangian_tendency_of_air_pressure"]
+    terminal_fall_speed_12 = contrail_12["terminal_fall_speed"]
 
-    longitude_t2, latitude_t2 = geo.advect_horizontal(
-        longitude=longitude,
-        latitude=latitude,
+    u_wind = 0.5 * (u_wind_1 + u_wind_12)
+    v_wind = 0.5 * (v_wind_1 + v_wind_12)
+    vertical_velocity = 0.5 * (vertical_velocity_1 + vertical_velocity_12)
+    longitude_2, latitude_2 = geo.advect_horizontal(
+        longitude=longitude_1,
+        latitude=latitude_1,
         u_wind=u_wind,
         v_wind=v_wind,
         dt=dt,
     )
-    level_t2 = geo.advect_level(level, vertical_velocity, rho_air, terminal_fall_speed, dt)
-    altitude_t2 = units.pl_to_m(level_t2)
+    level_2 = geo.advect_level(
+        level=level_1,
+        vertical_velocity=vertical_velocity,
+        rho_air=rho_air_1,
+        terminal_fall_speed=terminal_fall_speed_12,
+        dt=dt,
+    )
+    altitude_2 = units.pl_to_m(level_2)
+
+    # Using the _2 convention for post-advection data
+    index = contrail_1["index"]
+    formation_time = contrail_1["formation_time"]
+    time_2 = time_1 + dt
+    age_2 = age_1 + dt
 
     data = {
-        "index": index_t2,
-        "longitude": longitude_t2,
-        "latitude": latitude_t2,
-        "level": level_t2,
-        "air_pressure": 100.0 * level_t2,
-        "altitude": altitude_t2,
-        "time": time_t2,
+        "index": index,
+        "longitude": longitude_2,
+        "latitude": latitude_2,
+        "level": level_2,
+        "air_pressure": 100.0 * level_2,
+        "altitude": altitude_2,
+        "time": time_2,
         "formation_time": formation_time,
-        "age": age_t2,
-        **_get_uncertainty_params(contrail),
+        "age": age_2,
+        **_get_uncertainty_params(contrail_1),
     }
 
     if dt_tail is None or dt_head is None:
-        assert _is_segment_free_mode(contrail)
+        assert _is_segment_free_mode(contrail_1)
         assert dt_tail is None
         assert dt_head is None
-        return GeoVectorDataset._from_fastpath(data, attrs=contrail.attrs).copy()
+        return GeoVectorDataset._from_fastpath(data, attrs=contrail_1.attrs).copy()
 
-    longitude_head = contrail["longitude_head"]
-    latitude_head = contrail["latitude_head"]
-    longitude_tail = contrail["longitude_tail"]
-    latitude_tail = contrail["latitude_tail"]
-    u_wind_head = contrail["eastward_wind_head"]
-    v_wind_head = contrail["northward_wind_head"]
-    u_wind_tail = contrail["eastward_wind_tail"]
-    v_wind_tail = contrail["northward_wind_tail"]
+    longitude_head = contrail_1["longitude_head"]
+    latitude_head = contrail_1["latitude_head"]
+    longitude_tail = contrail_1["longitude_tail"]
+    latitude_tail = contrail_1["latitude_tail"]
+    u_wind_head_1 = contrail_1["eastward_wind_head"]
+    v_wind_head_1 = contrail_1["northward_wind_head"]
+    u_wind_tail_1 = contrail_1["eastward_wind_tail"]
+    v_wind_tail_1 = contrail_1["northward_wind_tail"]
+    u_wind_head_12 = contrail_12["eastward_wind_head"]
+    v_wind_head_12 = contrail_12["northward_wind_head"]
+    u_wind_tail_12 = contrail_12["eastward_wind_tail"]
+    v_wind_tail_12 = contrail_12["northward_wind_tail"]
+
+    u_wind_head = 0.5 * (u_wind_head_1 + u_wind_head_12)
+    v_wind_head = 0.5 * (v_wind_head_1 + v_wind_head_12)
+    u_wind_tail = 0.5 * (u_wind_tail_1 + u_wind_tail_12)
+    v_wind_tail = 0.5 * (v_wind_tail_1 + v_wind_tail_12)
 
     longitude_head_t2, latitude_head_t2 = geo.advect_horizontal(
         longitude=longitude_head,
@@ -2085,7 +2130,7 @@ def advect(
         lats1=latitude_tail_t2,
     )
 
-    head_tail_dt_t2 = np.full(contrail.size, np.timedelta64(0, "ns"))  # trivial
+    head_tail_dt_t2 = np.full(contrail_1.size, np.timedelta64(0, "ns"))  # trivial
 
     data["longitude_head"] = longitude_head_t2
     data["latitude_head"] = latitude_head_t2
@@ -2094,7 +2139,7 @@ def advect(
     data["segment_length"] = segment_length_t2
     data["head_tail_dt"] = head_tail_dt_t2
 
-    return GeoVectorDataset._from_fastpath(data, attrs=contrail.attrs).copy()
+    return GeoVectorDataset._from_fastpath(data, attrs=contrail_1.attrs).copy()
 
 
 def _aggregate_ef_summary(vector_list: list[VectorDataset]) -> VectorDataset | None: